Function (Bash/fish): cdb - Change Directory Below
Inspired by the DOS tool Directory Maven (created by Kent Briggs), cdb
is a simple (non-indexing) replica that issues a real-time find to get
the intended directory based on search term and then change to it.

Due to the real-time nature, an entire filesystem would be too resource
intensive to return a timely search (my home PC stores over two million
directories), so the compromise is to only search below the current
working directory (CWD).  Also, hidden directories (beginning with a
dot) are excluded in the find process.

After the initial concept was created (with a shim app made with Pascal),
with help the Bash version was simplified, and then I attempted to to the
same with the fish version.  Reaching out again I was given an example to
fill in the gaps.  As a result the function looks a bit different based
on which shell is being used.  Maybe I will clean that up later on.

Special thanks to:
Oliver Doepner (odoepner)
Kurtis Rader (krader1961)