Function (Bash/fish): clihelp - Shell Terminal Help + a Keybind or Two
Often I am forgetting what shortcuts are available to improve my CLI
experience.  So why not a popup bound to a key combo (in this case
CTRL+H) that displays them without disturbing the command entered (if
any)?

There is currently a bug with fish's implementation that the prompt may
not always show (even though it is forced), or at least to completion.
This may be the first time in this project that Bash easily does
exactly as intended (not that I am trying to make a bias).

Many tiny bits and pieces gathered for this one to judge any special
thanks, except for Fabian Homborg (faho) for the CTRL+C keybinding code
to add back to fish (clears/cancels current command).