Function (Bash/fish): mkd - Create and Enter into New Directory
I remember seeing something in DOS about recursive creation and entry
of directories many layers deep, did not realize that mkdir has an
argument (-p) that handles this part already.

This will create a directory, or series of directories inside of each
other, and then enter all the way down to the lowest branch.

Special thanks to:
Geoffroy Warin (geowarin)