Functions (Bash/fish): dirtoday, dirdays, cptoday, cpdays, dirrdays Identifies and/or copies recent files.
In part building on 'dirw', a large caveat of directory listings is the
lack of ability to filter the output in ways other than the file name.

These four functions either list or copy files in the current working
directory based on whether they were last modified today, or X (as a
parameter) number of days.  For 'cpdays' the parameter order is:

cpdays <X> <Destination>

'dirrdays' is the odd one out, as it does not have column control.  The
reason for this is that this command gets a recursive listing from the
current working directory, down (ignoring "hidden" files/directories).

I should note that 'dirdays 1' does not equal 'dirtoday', in that the
former is selecting files from the past 24 hours while the latter is
only selecting files modified since midnight.

Special thanks to:
Roberto Tomas
Stéphane Chazelas