Set include for .bash_extras (user-specific config)
While deploying CPC on a few systems, I found myself going back to edit
the .bashrc file to add back local tweaks (namely user-specific paths).

This commit will allow the local tweaks to be stored in a separate
file (.bash_extras), which will not get overwritten.  In the case that
the file does not exist, Bash will not attempt to call/include.