Script: wakessh - Invoke WOL (if necessary) and SSH in
Sometimes things make more sense when not a function.  Don't have to
create two variations for the same result, allows for more portability
being two examples.  WakeSSH would be the first of many scripts I
intend to add to CLIPC.

This particular script (in an earlier implementation) saved effort for
accessing a basement server.  Wake-On-LAN (WOL) fires it up, and then
the script pings approximately every twenty seconds for five minutes.
If there is a successful ping within that time, and SSH connection is
established.  The downside was that the MAC and IP addresses were
hard-coded into the script.

With this version, it relies on a configuration file (not to worry, in
the absence of one a template will be made) to define profiles (yes,
this script will allow for more than one configuration) including the
default (each profile can detail MAC, Host/IP, username, SSH port).

Any remaining options can be passed straight through to SSH (currently
excluding suffix/tail options, i.e. a particular command) as WakeSSH
parameters do not share the same case-sensitive letters as SSH (with
exception to the -h / --help option).

Special thanks to anyone who provided one of many hints along the way.
One of which was a commenter named Ahmed attached to a nixCraft post.