Bugfixes (Bash/fish): cdfl; (fish): act-core/dir-core
Multiple issuues afoot.

An easy win with the dependencies to `dirlast/dirprev/actlast/actprev`
(affecting fish only) as a test for number of parameters available was
not done in time since the move to the two core functions.  Adjusted
the initial expectations via minimal count and subsequent test to be
less forgiving.

Affecting both Bash and fish, `cdfl` was not only avoiding entry into
symlink directories, but a parent with the same initial as the child,
all matching the search criteria, will repeat the parent in the next
round of search results.

For example, `cdfl dmm` will return "device/motorola/msm8916-common" as
well as "device/motorola".  This was caused by the (partial) name
matching in the `find` results.  It would have been nice to use the
"-wholename" parameter instead, but there are serious restrictions with
that route, including any limitation of forward-slash separators that
deemed incompatible with the current code style.  The solution (or
kludge) has been to either trim out the list of nonsense results (fish)
or to transfer only the relevant results into a temporary array and
then flip the results back because `unset` is an upset (Bash).

Regarding the symlink issue, with the addition of the "-L" parameter
into the `find` command (follow symlinks), it made a directive or two
pointless ("-type l" lies when -L is present, taking away that and the
"-o" logical or) and caused some happy cleanup.